feat(swap): refactor SwapWidget
- moves connectbutton to shared component dir

- separates component + styling logic into separate files
